fix : parse処理をparseファイルにまとめた
